# Handle AMP etc. shortcodes