--- a/hugolib/hugo_sites_build.go
+++ b/hugolib/hugo_sites_build.go
@@ -213,7 +213,7 @@ func (h *HugoSites) render(config *BuildCfg) error {
 		s.initRenderFormats()
 		for i, rf := range s.renderFormats {
 			s.rc = &siteRenderingContext{Format: rf}
-			s.preparePagesForRender(i, config)
+			s.preparePagesForRender(config)
 
 			if !config.SkipRender {
 				if err := s.render(i); err != nil {