Fixes #3397