@bep 2017-05-09
For the curious; this is working fairly well at the moment, but it is low on tests (will fix ...) and there probably are some cases in the reload department I have not covered.

It is slightly slower than the current master, but slightly better than expected:

```
▶  benchcmp -best /tmp/1.bench /tmp/2.bench
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     11285502934     12559606511     +11.29%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     53325560       53764018       +0.82%

benchmark           old bytes       new bytes       delta
BenchmarkHugo-4     11372957440     11421035848     +0.42%
```