@bep 2017-05-10
I think I have fixed the main inefficiency with the new way of "doing things", and now it seems to be about as fast as before:

```
▶ benchcmp -best /tmp/1.bench /tmp/2.bench
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     11113985599     11488275340     +3.37%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     53325524       53658003       +0.62%

benchmark           old bytes       new bytes       delta
BenchmarkHugo-4     11371521072     11386936408     +0.14%
```