@bogem 2017-05-13
Hopefully #3443 will increase the performance.